pyproject.toml
[tool.poetry]
name = "repl_python3_BotDiscordPython"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
wikipedia = "^1.4.0"
discord = "^1.0.1"
flask-login = "^0.5.0"
colorama = "^0.4.4"
requests-futures = "^1.0.0"
torrequest = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
